rename to acpSessionRegistry
