+++ ui/goose2/src/shared/api/__tests__/acpNotificationHandler.test.ts
@@ -11,7 +11,7 @@ import {
   handleSessionNotification,
   setActiveMessageId,
 } from "../acpNotificationHandler";
-import { registerSession } from "../acpSessionTracker";
+import { registerSession } from "../acpSessionRegistry";
 
 function createMcpAppPayload(): McpAppPayload {
   return {
