+++ ui/goose2/src/shared/api/acpSessionRegistry.ts
